Create basic loop to display books
Created the displayLibrary function to iterate through the books and
properties of each book to display in divs on the screen. Added bare
CSS to set the library div to grid display. Future edits will add more
styling to make it look like a real book display instead of plain text.