Create and script add book form
Created new book form element in html. Set display to hidden and added
button that shows the form in the window. Added button that accepts
the input values and uses them to add the book details to the library
array. Button also adds the new book to the window as a div element so
that it appears with the pre-existing books on the page.