Create button to toggle read status
Wrote script to add a button when the books are added to the page that
can be used to toggle the read status of the books. Added toggleRead to
the book prototype to toggle between the true and false state on the
read property. Added script to change the text content of the read
status on each book when the state is toggled. Wrapping up the project
here because the focus was on the script, might come back to style it
another time.